Task  integrate runner (#9)
Patrika runner is now integrated into the Patrika package. This package now provides a command line interface which accepts a configuration json file along with the template JavaScript to generate your static site. Details are in the updated README.

Additionally, we are no longer publishing this to the github package repository, opting instead to publish it on the akshatmedia package repository (https://warehouse.akshatmedia.com).