We finally have the template correctly rerendering when the rendering code is updated. This was broken ever since we switched from CJS to ESM, because we could evict stuff from require.cache when using CJS. We _could_ do it by creating a require with ESMs, but that require didn't seem to include local imports, so it didn't work for our purpose and I'm not entirely sure why.
We ended up solving this issue by spinning up a new Worker thread every time we need to rebuild. The worker thread loads the template, and because it's a new worker every time, we get a fresh template.

We need to do some tidying up before we merge into master. I logged a new task in our kanpilot board about switching from template.js to config.json. The task has more details including a checklist and a deeper explanation of the problem.

Additionally,
- unused vars are now a warning instead of an error
- moved copyStaticAssets to the runner directory, since that is where it will be called from.