Use workers for better live rendering (#11)
The previous watch / live rendering functionality had an issue where it would work fine for the content (markdown) but not work well for the template (javascript). That was happening because I haven't been able to effectively reload modules (thanks ESM).

To get around that issue, we now perform the actual rendering by spinning up a new worker thread which in-turn loads the template. this way we load the latest template every time.

This commit also includes

Log improvements and kludges.
Fix for css file output (still debating whether to keep style building in Patrika or not, but it is handy for the moment).
Ability to copy over static assets.